% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{if_add}
\alias{if_add}
\alias{if_mul}
\alias{if_scalar_mul}
\alias{if_pow}
\alias{if_einstein_add}
\alias{if_einstein_mul}
\alias{if_einstein_scalar_mul}
\alias{if_einstein_pow}
\title{Intuitionistic fuzzy arithmetic (Xu operations)}
\usage{
if_add(a, b)

if_mul(a, b)

if_scalar_mul(lambda, a)

if_pow(a, lambda)

if_einstein_add(a, b)

if_einstein_mul(a, b)

if_einstein_scalar_mul(lambda, a)

if_einstein_pow(a, lambda)
}
\arguments{
\item{a, b}{objects coercible to \code{ifv} (recycled to common length).}

\item{lambda}{positive numeric scalar or vector.}
}
\value{
An \code{ifv} object.
}
\description{
Addition, multiplication, scalar multiplication and power for IFVs,
following Xu (2007): for \eqn{a=(\mu_a,\nu_a)}, \eqn{b=(\mu_b,\nu_b)},
\eqn{\lambda>0}:
\deqn{a\oplus b=(\mu_a+\mu_b-\mu_a\mu_b,\ \nu_a\nu_b)}
\deqn{a\otimes b=(\mu_a\mu_b,\ \nu_a+\nu_b-\nu_a\nu_b)}
\deqn{\lambda a=(1-(1-\mu_a)^\lambda,\ \nu_a^\lambda)}
\deqn{a^\lambda=(\mu_a^\lambda,\ 1-(1-\nu_a)^\lambda)}
}
\details{
Einstein variants use Einstein t-norm / t-conorm:
\deqn{a\oplus_\epsilon b=
  \left(\frac{\mu_a+\mu_b}{1+\mu_a\mu_b},
        \frac{\nu_a\nu_b}{1+(1-\nu_a)(1-\nu_b)}\right)}
\deqn{a\otimes_\epsilon b=
  \left(\frac{\mu_a\mu_b}{1+(1-\mu_a)(1-\mu_b)},
        \frac{\nu_a+\nu_b}{1+\nu_a\nu_b}\right)}
}
\examples{
a <- ifv(0.6, 0.3); b <- ifv(0.5, 0.3)
if_add(a, b); if_mul(a, b); if_scalar_mul(2, a); if_pow(a, 2)
}
\references{
Xu, Z. (2007). Intuitionistic fuzzy aggregation operators.
\emph{IEEE Trans. Fuzzy Systems}, 15(6), 1179-1187.
}
