% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifv.R
\name{if_check}
\alias{if_check}
\title{Validate IFVs}
\usage{
if_check(x, tol = 1e-09)
}
\arguments{
\item{x}{object coercible to \code{ifv}.}

\item{tol}{numeric tolerance.}
}
\value{
invisibly \code{TRUE}; errors if invalid.
}
\description{
Validate IFVs
}
