% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifv.R
\name{if_comp}
\alias{if_comp}
\title{Complement of an IFV}
\usage{
if_comp(x)
}
\arguments{
\item{x}{object coercible to \code{ifv}.}
}
\value{
\code{ifv} with mu/nu swapped.
}
\description{
Complement of an IFV
}
\examples{
if_comp(ifv(0.6, 0.2))
}
