% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifv.R
\name{if_pi}
\alias{if_pi}
\title{Hesitation degree}
\usage{
if_pi(x)
}
\arguments{
\item{x}{an \code{ifv} object (or coercible).}
}
\value{
numeric vector of \eqn{\pi = 1 - \mu - \nu}.
}
\description{
Hesitation degree
}
