% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifv.R
\name{is_ifv}
\alias{is_ifv}
\alias{as_ifv}
\title{Test / coerce intuitionistic fuzzy values}
\usage{
is_ifv(x)

as_ifv(x, ...)
}
\arguments{
\item{x}{object to test or coerce.}

\item{...}{ignored.}
}
\value{
\code{is_ifv()} logical; \code{as_ifv()} an \code{ifv} object.
}
\description{
Test / coerce intuitionistic fuzzy values
}
